requirement.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
pymysql
python-dotenv
